+++ azext_iot/tests/central/__init__.py
@@ -40,6 +40,9 @@ class CentralLiveScenarioTest(CaptureOutputLiveScenarioTest):
         )
         return super().cmd(command, checks=checks, expect_failure=expect_failure)
 
+    def cmd_withoutParams(self, command, checks=None, expect_failure=False):
+        return super().cmd(command, checks=checks, expect_failure=expect_failure)
+
     def _create_device(self, api_version, **kwargs) -> Tuple[str, str]:
         """
         kwargs: